@@ -1,12 +1,18 @@
 {
   "name": "shopify-admin-api-typings",
-  "version": "1.2.2",
+  "version": "1.3.0",
   "description": "Typings for Shopify Admin API",
+  "module": "esnext",
+  "type": "module",
   "author": {
     "name": "Carl-Johan Kihl",
     "email": "carljohan.kihl@gmail.com",
     "url": "https://carljohan.me"
   },
+  "contributors": [
+    "Carl-Johan Kihl",
+    "Riyad Shauk"
+  ],
   "repository": "github:caki0915/shopify-admin-api-typings",
   "bugs": {
     "url": "https://github.com/caki0915/shopify-admin-api-typings/issues"
@@ -19,12 +25,12 @@
   "main": "index.js",
   "types": "index.d.ts",
   "devDependencies": {
-    "dotenv": "8.2.0",
-    "graphql": "14.6.0",
-    "graphql-request": "1.8.2",
-    "graphql-schema-typescript": "1.3.2",
-    "lodash": "4.17.15",
-    "typescript": "3.9.3"
+    "dotenv": "10.0.0",
+    "graphql": "15.6.1",
+    "graphql-request": "3.5.0",
+    "graphql-schema-typescript": "1.5.2",
+    "lodash": "4.17.21",
+    "typescript": "4.4.3"
   },
   "peerDependencies": {
     "graphql": ">=14"
